requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20

# GCP Libraries
google-cloud-storage==2.19.0
google-cloud-firestore==2.20.0
google-generativeai==0.8.3

# Video Processing
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
